requirements.txt

fastapi
uvicorn
python-barcode
pillow
